Updated cgmanalysis after adding new data. Caught major bug.
BUG - cgmvariables was returning the data in the order of the files it read in, which wasn't the actual order due to how R sorts files (i.e. test1.csv, test10.csv, test2.csv).

This was not corrected/resorted when did the data analysis which resulted in the very high errors seen for CGMAnalysis.